大功告成，全部测试通过

+++ src/main/java/com/hellozjf/test/roboticrover/domainobject/Place.java
@@ -24,11 +24,11 @@ public class Place {
 
     /**
      * 角度信息，0表示正东，90表示正北，180表示正西，270表示正南，360表示正东……
-     * 移动的时候，x = x + cos(angle)，y = y + sin(angle)
+     * 移动的时候，x = x + cos(rad(angle))，y = y + sin(rad(angle))
      */
-    private Double angle;
+    private Integer angle;
 
-    public Place(Integer x, Integer y, Double angle) {
+    public Place(Integer x, Integer y, Integer angle) {
         this.x = x;
         this.y = y;
         this.angle = angle;
